chore: move SSH to its own submodule
