+++ src/controllers/ssh/mod.rs
@@ -1,0 +1,4 @@
+pub mod authentication;
+pub mod keys;
+
+pub use authentication::authenticate;
